+++ common/src/main/java/com/kwwsyk/endinv/common/NbtAttachment.java
@@ -2,8 +2,10 @@ package com.kwwsyk.endinv.common;
 
 import net.minecraft.world.entity.player.Player;
 
-public interface NbtAttachment<T> {
+import javax.annotation.Nullable;
 
+public interface NbtAttachment<T> {
+    @Nullable
     T getWith(Player player);
 
     void setTo(Player player, T t);