+++ neoforge/src/main/java/com/kwwsyk/endinv/neoforge/events/PlayerEvents.java
@@ -1,19 +1,12 @@
 package com.kwwsyk.endinv.neoforge.events;
 
 import com.kwwsyk.endinv.common.ModInfo;
-import com.kwwsyk.endinv.common.ModRegistries;
-import com.kwwsyk.endinv.common.ServerLevelEndInv;
-import com.kwwsyk.endinv.common.network.payloads.toClient.EndInvContent;
-import com.kwwsyk.endinv.common.network.payloads.toClient.EndInvMetadata;
-import com.kwwsyk.endinv.common.network.payloads.toClient.MenuAttachabilityPayload;
-import com.kwwsyk.endinv.common.options.ContentTransferMode;
-import com.kwwsyk.endinv.neoforge.options.ServerConfig;
+import com.kwwsyk.endinv.common.event.PlayerEventHelper;
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 import net.neoforged.neoforge.event.tick.PlayerTickEvent;
-import net.neoforged.neoforge.network.PacketDistributor;
 
 @EventBusSubscriber(modid = ModInfo.MOD_ID)
 public class PlayerEvents {
@@ -23,23 +16,7 @@ public class PlayerEvents {
     public static void tick(PlayerTickEvent.Post event){
         if(event.getEntity() instanceof ServerPlayer serverPlayer){
             if(tickRefresh) {
-                PacketDistributor.sendToPlayer(serverPlayer, ModRegistries.NbtAttachments.getSyncedConfig().computeIfAbsent(serverPlayer));
-                // Send effective menu attachability
-                var serverCfg = ModInfo.getServerConfig();
-                var menuCfg = serverCfg.specifiedMenuAttachability().get();
-                boolean defaultAttach = serverCfg.enableAttaching().get();
-                PacketDistributor.sendToPlayer(serverPlayer,
-                        new MenuAttachabilityPayload(
-                                defaultAttach,
-                                menuCfg.isInventoryAttachable(),
-                                menuCfg.getConfigs()
-                        ));
-                if(ServerConfig.CONFIG.TRANSFER_MODE.get()== ContentTransferMode.ALL){
-                    ServerLevelEndInv.getEndInvForPlayer(serverPlayer).ifPresent(endInv -> {
-                        PacketDistributor.sendToPlayer(serverPlayer,new EndInvContent(endInv.getItemMap()));
-                        PacketDistributor.sendToPlayer(serverPlayer,EndInvMetadata.getWith(endInv));
-                    });
-                }
+                PlayerEventHelper.sendData(serverPlayer);
                 tickRefresh = false ;
             }
         }